+++ CHANGELOG.md
@@ -96,6 +96,7 @@
 * FEATURE: add Dell X-series model (@RobbFromIT)
 * FEATURE: add privilege escalation to the cumulus model (@user4574)
 * FEATURE: add adtran model (@CFUJoshWeepie)
+* FEATURE: netscaler, add multiple partition support (@smallsam)
 * FEATURE: add firebrick model (@lewisvive)
 * BUGFIX: netgear telnet password prompt not detected
 * BUGFIX: xos model should not modify config on legacy Extreme Networks devices (@sq9mev)